requirements.txt
pandas>=1.5.0
requests>=2.28.0
urllib3>=1.26.0
python-dotenv>=1.0.0
flask>=2.3.0
